Fixes for the PS4 and PSP2 ports
I still feel like the game is running too slowly but perhaps it is running as intended. I've never played doom and frankly don't really care for it at all.

The PS4 port now shares pad controls with the PSP2 version.
I've cleaned the PS4 code a bit, but it is still far too messy.
In the future, I'd like to stop using libSceGnm completely, because I don't think it is required.